tools/mqtt_logger/pyproject.toml
[project]
name = "multigeiger-mqtt-logger"
version = "1.0.0"
description = "MQTT data logger for MultiGeiger radiation detector"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "paho-mqtt>=1.6.1",
    "python-dotenv>=1.0.0",
    "matplotlib>=3.5.0",
]

[project.scripts]
mqtt-logger = "mqtt_logger:main"
